+++ src/components/pages/common/sign-message/types.tsx
@@ -1,0 +1,7 @@
+import { TDropStep, TMultiscanStep } from 'types'
+
+type TProps = {
+  onSubmit: () => void
+}
+
+export default TProps